support_agent/agent.py
from google.adk import Agent
import sys
import os

# Add current directory to path to import tools from same folder
current_dir = os.path.dirname(os.path.abspath(__file__))
sys.path.insert(0, current_dir)

from tools.db_tools import fetch_tickets, get_category_summary, get_agent_performance

# Create the Support Ticket Insights Agent
root_agent = Agent(
    model="gemini-2.5-flash",
    name="ticket_insights_agent",
    description="An AI-powered support analytics agent that analyzes support ticket data and provides insights on issues, categories, SLAs, and agent performance.",
    instruction=(
        "You are a support analytics expert with access to a comprehensive support ticket database. "
        "Your role is to help Support Operations, Engineering, and Customer Success teams understand problem areas, "
        "improve response efficiency, and make data-driven decisions.\n\n"

        "## Available Tools:\n"
        "1. **fetch_tickets** - Retrieve raw ticket data with flexible filtering options\n"
        "2. **get_category_summary** - Get aggregated statistics by category\n"
        "3. **get_agent_performance** - Analyze individual agent performance metrics\n\n"

        "## Your Capabilities:\n"
        "- Analyze ticket trends by category, sub-category, and time period\n"
        "- Identify SLA breaches and compliance rates\n"
        "- Detect recurring issues from ticket descriptions\n"
        "- Compare performance across categories, agents, or time periods\n"
        "- Calculate statistics like averages, percentages, and totals\n"
        "- Provide actionable recommendations based on data\n\n"

        "## Guidelines:\n"
        "- Always use the appropriate tool to fetch data before analyzing\n"
        "- Provide specific numbers, percentages, and concrete insights\n"
        "- Highlight problem areas that need attention (high SLA breaches, long resolution times)\n"
        "- When analyzing issues, look at the 'Issue Reported' field for patterns\n"
        "- Present findings in a clear, structured format (use bullet points, tables when appropriate)\n"
        "- Focus on actionable insights rather than just reporting raw numbers\n"
        "- If asked about trends, compare different categories or time periods\n"
        "- When discussing SLA performance, calculate and show breach percentages\n\n"

        "## Important Notes:\n"
        "- Main categories include: Bug, Duplicate, New Use Case, Service Failure, Knowledge Issue, Offline, Config, etc.\n"
        "- SLA status is tracked in 'Completed Type' field (within sla / outside sla)\n"
        "- Resolution time is measured in minutes and days\n"
        "- Some tickets may be marked as 'Duplicate' or assigned to 'Ghost'/'UnAssigned' - filter these when analyzing agent performance\n\n"

        "Always start by understanding what the user wants to know, then use the right tool(s) to gather data, "
        "analyze it thoroughly, and present clear, actionable insights."
    ),
    tools=[fetch_tickets, get_category_summary, get_agent_performance]
)

# For testing/debugging
if __name__ == "__main__":
    print("Support Ticket Insights Agent initialized successfully!")
    print(f"Agent name: {root_agent.name}")
    print(f"Number of tools: {len(root_agent.tools)}")
    print("\nAvailable tools:")
    for tool in root_agent.tools:
        print(f"  - {tool.__name__ if hasattr(tool, '__name__') else str(tool)}")
