support_agent/requirements.txt
pandas
python-dotenv
requests
google-generativeai
google-adk
